// src/redux/Modules/mapsPolygons/maps.d.ts
declare namespace Farms {
  export interface FarmsData {
    farmsDataArray: Array<any>;
    singleFarm: any;
    filesList: any;
    cropName: string;
    farmName: string;
    taskFilterOpen: any;
    summaryFarmTitle: string;
    summaryCropTitle: string;
    polygonCoords: any;
    searchLocation: any;
  }
}

export { Farms };